+++ Runtime/Api/Session.cs
@@ -1,17 +1,17 @@
-﻿// Copyright (c) 2022 AccelByte Inc. All Rights Reserved.
+﻿// Copyright (c) 2022 - 2024 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
 using AccelByte.Core;
 using AccelByte.Models;
+using AccelByte.Api.Interface;
 using Newtonsoft.Json.Linq;
-using System;
 using System.Collections.Generic;
 using UnityEngine.Assertions;
 
 namespace AccelByte.Api
 {
-    public class Session : WrapperBase
+    public class Session : WrapperBase, IClientSession
     {
         private readonly SessionApi sessionApi;
         private readonly ISession session;
@@ -30,14 +30,6 @@ namespace AccelByte.Api
         }
 
         #region PartySession
-
-        /// <summary>
-        /// Get Party details with partyId.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID.</param>
-        /// <param name="callback">
-        /// Returns a Result that contain SessionV2PartySession via callback when completed.
-        /// </param>
         public void GetPartyDetails(string partyId, ResultCallback<SessionV2PartySession> callback)
         {
             Report.GetFunctionLog(GetType().Name);
@@ -59,17 +51,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Update(Overwrite) all fields of a party.
-        /// Note: Join type can only be updated by the party's leader.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID.</param>
-        /// <param name="request">
-        /// Compiled request with the updating details of the party session
-        /// </param>
-        /// <param name="callback">
-        /// Returns a Result that contain SessionV2PartySession via callback when completed.
-        /// </param>
         public void UpdateParty(string partyId, SessionV2PartySessionUpdateRequest request,
             ResultCallback<SessionV2PartySession> callback)
         {
@@ -96,17 +77,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Update(Partially) specified field(s) of a party.
-        /// Note: Join type can only be updated by the party's leader.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID.</param>
-        /// <param name="request">
-        /// Compiled request with the updating details of the party session
-        /// </param>
-        /// <param name="callback">
-        /// Returns a Result that contain SessionV2PartySession via callback when completed.
-        /// </param>
         public void PatchUpdateParty(string partyId, SessionV2PartySessionUpdateRequest request,
             ResultCallback<SessionV2PartySession> callback)
         {
@@ -132,12 +102,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Invite a user to a party.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID</param>
-        /// <param name="userId">Targeted user ID</param>
-        /// <param name="callback">Returns a Result with status code</param>
         public void InviteUserToParty(string partyId, string userId,
             ResultCallback callback)
         {
@@ -174,12 +138,6 @@ namespace AccelByte.Api
                     }));
         }        
         
-        /// <summary>
-        /// Promotes a party member to be a party leader. Only leader can promote a new leader.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID</param>
-        /// <param name="leaderId">Targeted user ID of new leader</param>
-        /// <param name="callback">Returns a Result that contain SessionV2PartySession via callback when completed</param>
         public void PromoteUserToPartyLeader(string partyId, string leaderId,
             ResultCallback<SessionV2PartySession> callback)
         {
@@ -216,13 +174,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Join a party.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID.</param>
-        /// <param name="callback">
-        /// Returns a Result that contain SessionV2PartySession via callback when completed.
-        /// </param>
         public void JoinParty(string partyId, ResultCallback<SessionV2PartySession> callback)
         {
             Report.GetFunctionLog(GetType().Name);
@@ -251,11 +202,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Leave a Party.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID.</param>
-        /// <param name="callback">Returns a Result with status code</param>
         public void LeaveParty(string partyId, ResultCallback callback)
         {
             Report.GetFunctionLog(GetType().Name);
@@ -284,11 +230,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Reject a party invitation.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID.</param>
-        /// <param name="callback">Returns a Result with status code</param>
         public void RejectPartyInvitation(string partyId, ResultCallback callback)
         {
             Report.GetFunctionLog(GetType().Name);
@@ -310,15 +251,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Kick a player from a party.
-        /// Requires invoker to be the party leader.
-        /// </summary>
-        /// <param name="partyId">Targeted party ID.</param>
-        /// <param name="userId">Targeted user's ID.</param>
-        /// <param name="callback">
-        /// Returns a Result that contain SessionV2PartySessionKickResponse via callback when completed.
-        /// </param>
         public void KickUserFromParty(string partyId, string userId,
             ResultCallback<SessionV2PartySessionKickResponse> callback)
         {
@@ -354,15 +286,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Create a Party and assign invoker as leader.
-        /// </summary>
-        /// <param name="request">
-        /// Compiled request with details of the creating party session
-        /// </param>
-        /// <param name="callback">
-        /// Returns a Result that contain SessionV2PartySession via callback when completed.
-        /// </param>
         public void CreateParty(SessionV2PartySessionCreateRequest request,
             ResultCallback<SessionV2PartySession> callback)
         {
@@ -388,12 +311,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Get all parties of a user(invoker).
-        /// </summary>
-        /// <param name="callback">
-        /// Returns a paginated result of SessionV2PartySession via callback when completed.
-        /// </param>
         public void GetUserParties(ResultCallback<PaginatedResponse<SessionV2PartySession>> callback)
         {
             Report.GetFunctionLog(GetType().Name);
@@ -409,11 +326,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Join a party using a code.
-        /// </summary>
-        /// <param name="code">Party code.</param>
-        /// <param name="callback">Returns a result of SessionV2PartySession via callback when completed.</param>
         public void JoinPartyByCode(string code
             , ResultCallback<SessionV2PartySession> callback)
         {
@@ -437,11 +349,6 @@ namespace AccelByte.Api
                 }));
         }
 
-        /// <summary>
-        /// Generate new code for a party.
-        /// </summary>
-        /// <param name="partyId">Party ID.</param>
-        /// <param name="callback">Returns a result of SessionV2PartySession via callback when completed.</param>
         public void GenerateNewPartyCode(string partyId
             , ResultCallback<SessionV2PartySession> callback)
         {
@@ -462,11 +369,6 @@ namespace AccelByte.Api
                 sessionApi.GenerateNewPartyCode(partyId, callback));
         }
 
-        /// <summary>
-        /// Revoke party code.
-        /// </summary>
-        /// <param name="sessionId">Party ID.</param>
-        /// <param name="callback">Returns a result of SessionV2PartySession via callback when completed.</param>
         public void RevokePartyCode(string sessionId
             , ResultCallback callback)
         {
@@ -491,13 +393,6 @@ namespace AccelByte.Api
 
         #region GameSession
 
-        /// <summary>
-        /// Create Game Session
-        /// </summary>
-        /// <param name="request">Details of creating game session.</param>
-        /// <param name="callback">
-        /// Returns a Result that contain SessionV2GameSession via callback when completed.
-        /// </param>
         public void CreateGameSession(SessionV2GameSessionCreateRequest request
             , ResultCallback<SessionV2GameSession> callback)
         {
@@ -524,14 +419,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Query Game Session
-        /// By default, API will return a list of available game sessions
-        /// </summary>
-        /// <param name="request">List of attributes to filter from available sessions</param>
-        /// <param name="callback">
-        /// Returns a paginated result of SessionV2GameSession via callback when completed.
-        /// </param>
         public void QueryGameSession(Dictionary<string, object> request
             , ResultCallback<PaginatedResponse<SessionV2GameSession>> callback)
         {
@@ -553,14 +440,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Get game session detail by podName.
-        /// Session service has several DSInformation status to track DS request to DSMC
-        /// </summary>
-        /// <param name="podName">Targeted game session's podName</param>
-        /// <param name="callback">
-        /// Returns a result of SessionV2GameSession via callback when completed.
-        /// </param>
         public void GetGameSessionDetailsByPodName(string podName
             , ResultCallback<SessionV2GameSession> callback)
         {
@@ -579,14 +458,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Get game session detail by sessionId.
-        /// Session service has several DSInformation status to track DS request to DSMC
-        /// </summary>
-        /// <param name="sessionId">Targeted game session's sessionId</param>
-        /// <param name="callback">
-        /// Returns a result of SessionV2GameSession via callback when completed.
-        /// </param>
         public void GetGameSessionDetailsBySessionId(string sessionId
             , ResultCallback<SessionV2GameSession> callback)
         {
@@ -609,13 +480,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Delete a game session.
-        /// </summary>
-        /// <param name="sessionId">Targeted game session's sessionId</param>
-        /// <param name="callback">
-        /// Returns a result via callback when completed.
-        /// </param>
         public void DeleteGameSession(string sessionId
             , ResultCallback callback)
         {
@@ -638,14 +502,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Patch updates a game session, only specified fields from game session data.
-        /// </summary>
-        /// <param name="sessionId">Targeted game session's sessionId</param>
-        /// <param name="request">Compiled request of updating game session details</param>
-        /// <param name="callback">
-        /// Returns a result of updated SessionV2GameSession via callback when completed.
-        /// </param>
         public void PatchGameSession(string sessionId, SessionV2GameSessionUpdateRequest request
             , ResultCallback<SessionV2GameSession> callback)
         {
@@ -670,14 +526,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Invite a user to a game session.
-        /// </summary>
-        /// <param name="sessionId">Targeted game session's sessionId</param>
-        /// <param name="userId">Targeted user's userId</param>
-        /// <param name="callback">
-        /// Returns a result of SessionV2GameSession via callback when completed.
-        /// </param>
         public void InviteUserToGameSession(string sessionId, string userId
             , ResultCallback callback)
         {
@@ -714,13 +562,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Join a game session.
-        /// </summary>
-        /// <param name="sessionId">Targeted game session's sessionId</param>
-        /// <param name="callback">
-        /// Returns a result of SessionV2GameSession via callback when completed.
-        /// </param>
         public void JoinGameSession(string sessionId
             , ResultCallback<SessionV2GameSession> callback)
         {
@@ -750,13 +591,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Leave a game session.
-        /// </summary>
-        /// <param name="sessionId">Targeted game session's sessionId</param>
-        /// <param name="callback">
-        /// Returns a result of SessionV2GameSession via callback when completed.
-        /// </param>
         public void LeaveGameSession(string sessionId
             , ResultCallback<SessionV2GameSession> callback)
         {
@@ -786,13 +620,6 @@ namespace AccelByte.Api
                     }));
         }
 
-        /// <summary>
-        /// Reject a game session invitation.
-        /// </summary>
-        /// <param name="sessionId">Targeted game session's sessionId</param>
-        /// <param name="callback">
-        /// Returns a result via callback when completed.
-        /// </param>
         public void RejectGameSessionInvitation(string sessionId
             , ResultCallback callback)
         {
@@ -815,18 +642,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Query user's game sessions.
-        /// By default, API will return a list of user's active game sessions (INVITED,JOINED,CONNECTED).
-        /// </summary>
-        /// <param name="statusFilter">Game session status to filter. Supported status: INVITED, JOINED, CONNECTED</param>
-        /// <param name="orderBy">Order result by specific attribute. Supported: createdAt (default), updatedAt</param>
-        /// <param name="sortDesc">
-        /// Is descending order of the result. Default is true. Ascending order if false.
-        /// </param>
-        /// <param name="callback">
-        ///  Returns a paginated result of SessionV2GameSession via callback when completed.
-        /// </param>
         public void GetUserGameSessions(SessionV2StatusFilter? statusFilter, SessionV2AttributeOrderBy? orderBy,
             bool? sortDesc, ResultCallback<PaginatedResponse<SessionV2GameSession>> callback)
         {
@@ -846,11 +661,6 @@ namespace AccelByte.Api
                     callback));
         }
 
-        /// <summary>
-        /// Join a game session using a code.
-        /// </summary>
-        /// <param name="code">Game session code.</param>
-        /// <param name="callback">Returns a result of SessionV2GameSession via callback when completed.</param>
         public void JoinGameSessionByCode(string code
             , ResultCallback<SessionV2GameSession> callback)
         {
@@ -874,11 +684,6 @@ namespace AccelByte.Api
                 }));
         }
 
-        /// <summary>
-        /// Generate new code for a game session.
-        /// </summary>
-        /// <param name="sessionId">Game session ID.</param>
-        /// <param name="callback">Returns a result of SessionV2GameSession via callback when completed.</param>
         public void GenerateNewGameSessionCode(string sessionId
             , ResultCallback<SessionV2GameSession> callback)
         {
@@ -899,11 +704,6 @@ namespace AccelByte.Api
                 sessionApi.GenerateNewGameSessionCode(sessionId, callback));
         }
 
-        /// <summary>
-        /// Revoke game session code.
-        /// </summary>
-        /// <param name="sessionId">Game session ID.</param>
-        /// <param name="callback">Returns a result of SessionV2GameSession via callback when completed.</param>
         public void RevokeGameSessionCode(string sessionId
             , ResultCallback callback)
         {
@@ -948,14 +748,6 @@ namespace AccelByte.Api
 
         #region SessionStorage
 
-        /// <summary>
-        /// Update leader's session storage data, can only be updated by current session leader.
-        /// This will overwrite leader storage data, if updating also provide latest leader storage.
-        /// To clear current leader storage data update with empty JObject.
-        /// </summary>
-        /// <param name="sessionId">The game or party session id.</param>
-        /// <param name="data">Data to update leader storage.</param>
-        /// <param name="callback">Callback will be called when completed.</param>
         public void UpdateLeaderStorage(string sessionId, JObject data, ResultCallback<JObject> callback)
         {
             Report.GetFunctionLog(GetType().Name);
@@ -975,14 +767,6 @@ namespace AccelByte.Api
                 sessionApi.UpdateLeaderStorage(sessionId, data, callback));
         }
 
-        /// <summary>
-        /// Update current user's session member storage data.
-        /// This will overwrite this user's storage data, if updating also provide latest user's storage.
-        /// To clear current user's storage data update with empty jsonObject.
-        /// </summary>
-        /// <param name="sessionId">The game or party session id.</param>
-        /// <param name="data">Data to update leader storage.</param>
-        /// <param name="callback">Callback will be called when completed.</param>
         public void UpdateMemberStorage(string sessionId, JObject data, ResultCallback<JObject> callback)
         {
             Report.GetFunctionLog(GetType().Name);